fix(api): silence C4244 on meshletCount
size_t to double is an implicit narrowing MSVC warns about at /W4; the
value is a small count, so the cast is lossless in practice.

+++ src/api/native_mesh_analysis.cpp
@@ -415,7 +415,7 @@ void initMeshAnalysis(ObjectBuilder& proto, HostClass& cls) {
             return mo.build();
         });
 
-        ev::setProperty(arr, "meshletCount", ev::fromDouble(meshlets.size()));
+        ev::setProperty(arr, "meshletCount", ev::fromDouble(static_cast<double>(meshlets.size())));
         ev::Persistent vBuf(makeUint32Array(allVertices.data(), allVertices.size()));
         ev::setProperty(arr, "vertices", vBuf.get());
         ev::Persistent tBuf(makeUint8Array(allTriangles.data(), allTriangles.size()));
